# week-4/api/openapi.yaml
openapi: 3.0.3
info:
  title: PaaS Product API
  version: "1.0.0"
  description: |
    Provisions and manages managed Postgres database instances running on
    STACKIT SKE, backed by the CloudNativePG operator. Each "instance"
    represents one full database product (one CNPG Cluster, made up of one
    or more Postgres pods internally).
...
security:
  - bearerAuth: []
paths:
...
  /v1/instances:
    post:
      tags: [Instances]
      summary: Create a new product instance
      description: >
        Creates a new managed Postgres instance, owned by the authenticated
        user. Returns immediately once the request is accepted and stored —
        the instance is not yet running at this point (phase will be
        `Provisioning`). Poll GET /v1/instances/{id} to check when it becomes
        `Healthy`.
      operationId: createInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceRequest'
            example:
              name: api-test-1
              instances: 3
              storageSize: 5Gi
      responses:
        '201':
          description: Instance accepted and being provisioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Instances]
      summary: List your product instances
      description: Returns every instance owned by the authenticated user, regardless of phase.
      operationId: listInstances
      responses:
        '200':
          description: List of instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Instance'
  /v1/instances/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The instance's unique name (same value provided at creation time).
        schema:
          type: string
        example: api-test-1
    get:
      tags: [Instances]
      summary: Get a single instance
      description: Returns the current state of one instance, including its live phase.
      operationId: getInstance
      responses:
        '200':
          description: Instance details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Instances]
      summary: Update a product instance
      description: >
        Partially updates a running instance. Only `instances`, `storageSize`,
        and `allowedIPs` can be changed after creation — name, storageClass,
        database, and username are fixed at creation time. Omitted fields are
        left unchanged; `storageSize` may only be increased, never decreased;
        `allowedIPs: []` removes external access entirely.
      operationId: updateInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInstanceRequest'
            example:
              instances: 3
      responses:
        '200':
          description: Updated instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Instances]
      summary: Delete a product instance
      description: >
        Deletes the instance. Deletion is asynchronous — the underlying pods,
        volumes, and secrets are torn down by the operator shortly after this
        call returns, not synchronously within it.
      operationId: deleteInstance
      responses:
        '202':
          description: Deletion accepted
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/instances/{id}/connection:
    parameters:
      - name: id
        in: path
        required: true
        description: The instance's unique name.
        schema:
          type: string
        example: api-test-1
    get:
      tags: [Instances]
      summary: Get connection and access data for an instance
      description: >
        Returns the host, port, database name, and credentials needed to
        connect to this instance. Only available once the instance has
        finished bootstrapping.
      operationId: getInstanceConnection
      responses:
        '200':
          description: Connection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionInfo'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Instance exists but hasn't finished provisioning yet — no credentials available
  /v1/instances/{id}/services:
    parameters:
      - name: id
        in: path
        required: true
        description: The instance's unique name.
        schema:
          type: string
        example: api-test-1
    get:
      tags: [Instances]
      summary: List the Postgres endpoints exposed for an instance
      description: >
        Returns the ClusterIP Services CNPG manages automatically for this
        instance — the read/write primary endpoint ("-rw"), the read-only
        replica endpoint ("-ro"), and the any-replica endpoint ("-r").
      operationId: listInstanceServices
      responses:
        '200':
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceInfo'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token returned by POST /v1/auth/register or POST /v1/auth/login.
  schemas:
...
    CreateInstanceRequest:
      type: object
      required: [name, instances, storageSize, storageClass, database, username]
      properties:
        name:
          type: string
          pattern: '^[a-z][a-z0-9-]{0,9}$'
          description: >
            Unique instance name. Must start with a lowercase letter, contain
            only lowercase letters/digits/hyphens, and be 11 characters or
            fewer (a hard limit of the underlying SKE platform).
          example: api-test-1
        instances:
          type: integer
          minimum: 1
          maximum: 5
          description: Number of Postgres pods (1 primary + N-1 replicas) backing this instance.
        storageSize:
          type: string
          example: "5Gi"
          description: Disk size per Postgres pod, in Kubernetes quantity notation.
        storageClass:
          type: string
          enum: [premium-perf4-stackit]
          description: Kubernetes StorageClass backing each pod's volume. Currently only one is offered.
          example: premium-perf4-stackit
        database:
          type: string
          pattern: '^[a-zA-Z_][a-zA-Z0-9_]{0,62}$'
          description: Name of the database created inside the instance at bootstrap time.
          example: appdb
        username:
          type: string
          pattern: '^[a-zA-Z_][a-zA-Z0-9_]{0,62}$'
          description: Owner user created for the bootstrap database — this is what you authenticate as.
          example: appuser
        allowedIPs:
          type: array
          items:
            type: string
          description: >
            IP networks in CIDR notation allowed to reach this instance from
            outside the cluster. If omitted or empty, the instance is not
            exposed externally — no public IP is provisioned, and it stays
            reachable only from inside the cluster.
          example: ["203.0.113.0/24", "198.51.100.42/32"]
    UpdateInstanceRequest:
      type: object
      description: >
        All fields optional — only the ones present are changed. At least
        one must be given.
      properties:
        instances:
          type: integer
          minimum: 1
          maximum: 5
          description: New desired pod count. CNPG handles both scale-up and scale-down.
        storageSize:
          type: string
          description: >
            New disk size per pod, in Kubernetes quantity notation. Must be
            strictly greater than the current size — shrinking a live volume
            isn't supported and is rejected with 400.
          example: "10Gi"
        allowedIPs:
          type: array
          items:
            type: string
          description: >
            Replaces the full set of IP networks in CIDR notation allowed to
            reach this instance externally. An empty array removes external
            access (deletes the dedicated LoadBalancer Service); omitting
            this field entirely leaves external access as-is.
          example: ["203.0.113.0/24"]
    Instance:
      type: object
      properties:
        id:
          type: string
          description: Unique instance identifier (matches the name given at creation).
        uid:
          type: string
          description: >
            Kubernetes object UID backing this instance — a stable,
            globally-unique identifier distinct from the name-based `id`.
            Informational only; not accepted as input anywhere.
          example: aab7b8d6-17a7-4187-a68d-92aa236e6a3a
        name:
          type: string
        phase:
          type: string
          enum: [Provisioning, Healthy, Degraded, Deleting]
          description: Current lifecycle state, as reported by the underlying database operator.
        instances:
          type: integer
          description: Desired number of Postgres pods.
        readyInstances:
          type: integer
          description: Number of Postgres pods currently ready and serving traffic.
        external:
          type: boolean
          description: >
            Whether this instance was created with allowedIPs and has a
            dedicated external LoadBalancer Service. Does not by itself mean
            the external IP has finished being assigned — poll the
            connection endpoint for that.
        allowedIPs:
          type: array
          items:
            type: string
          description: >
            The CIDR ranges currently allowed to reach this instance
            externally. Absent/empty when the instance is internal-only.
          example: ["203.0.113.0/24"]
        version:
          type: string
          description: >
            Postgres version tag of the image the operator is currently
            running. Empty until the operator has resolved and reported an
            image, which isn't yet the case on a freshly created instance.
          example: "16.4"
        storageSize:
          type: string
          description: Disk size per Postgres pod, as requested at creation time.
          example: "5Gi"
        createdAt:
          type: string
          format: date-time
    ServiceInfo:
      type: object
      properties:
        name:
          type: string
          example: api-test-1-rw
        type:
          type: string
          example: ClusterIP
        clusterIP:
          type: string
          example: 100.82.187.136
        externalIP:
          type: string
          description: Empty for the internal rw/ro/r Services — they're never externally exposed directly.
        ports:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: postgres
              port:
                type: integer
                example: 5432
              protocol:
                type: string
                example: TCP
    ConnectionInfo:
      type: object
      properties:
        host:
          type: string
          description: >
            Address for the instance's primary (read/write) endpoint. If the
            instance was created with allowedIPs, this is the external
            LoadBalancer IP; otherwise it's the in-cluster DNS name, reachable
            only from inside the cluster.
          example: api-test-1-rw.cnpg-system.svc.cluster.local
        port:
          type: integer
          example: 5432
        database:
          type: string
          example: appdb
        username:
          type: string
          example: appuser
        password:
          type: string
          description: Auto-generated at instance creation. Not recoverable if lost — the instance must be recreated.
...
    Error:
      type: object
      properties:
        message:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Instance not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'